move types, remove other files
