--- frontend/src/lib/models/container-registry.ts
@@ -1,11 +1,0 @@
-export interface ContainerRegistry {
-    id: string;
-    url: string;
-    username: string;
-    token: string;
-    description?: string;
-    insecure?: boolean;
-    enabled?: boolean;
-    createdAt?: string;
-    updatedAt?: string;
-}
