+++ frontend/src/lib/types/container-registry.type.ts
@@ -15,3 +15,15 @@ export interface ContainerRegistryUpdateDto {
 	insecure?: boolean;
 	enabled?: boolean;
 }
+
+export interface ContainerRegistry {
+	id: string;
+	url: string;
+	username: string;
+	token: string;
+	description?: string;
+	insecure?: boolean;
+	enabled?: boolean;
+	createdAt?: string;
+	updatedAt?: string;
+}
