+++ frontend/src/lib/types/docker-info.type.ts
