+++ frontend/src/lib/types/environment.type.ts
