+++ frontend/src/lib/types/system-stats.type.ts
